Replace CMOV with obliMove, avoid template specialization and potential misuse

+++ omap/interface/recoram_interface.cpp
@@ -13,7 +13,7 @@ ORAMBindingSingleton::ORAMBindingSingleton() {
   oram = nullptr;
 }
 
-void ORAMBindingSingleton::InitORAM(uint64_t size) {
+void ORAMBindingSingleton::InitORAM(uint32_t size) {
   // ASSERT(oram == nullptr);
   oram = (void*)(new RecursiveORAM<uint64_t, uint32_t>(size));
   ((RecursiveORAM<uint64_t, uint32_t>*)oram)->InitDefault(0);
