+++ omap/interface/recoram_interface.hpp
@@ -5,8 +5,7 @@ typedef void* oram_t;
 struct ORAMBindingSingleton {
   oram_t oram;
   ORAMBindingSingleton();
-  void InitORAM(uint64_t size);
+  void InitORAM(uint32_t size);
   void Write(uint32_t addr, uint64_t val);
   uint64_t Read(uint32_t addr);
 };
-
